Use static casts, compare signed/unsigned like types to reduce warnings.

+++ cpp/order/HexaticTranslational.cc
@@ -54,7 +54,7 @@ void Hexatic::compute(const freud::locality::NeighborList* nlist,
     computeGeneral(
         [this](const vec3<float>& delta) {
             const float theta_ij = std::atan2(delta.y, delta.x);
-            return std::exp(std::complex<float>(0, m_k * theta_ij));
+            return std::exp(std::complex<float>(0, static_cast<float>(m_k) * theta_ij));
         },
         nlist, points, qargs, false);
 }
